improve embedding plot
The annotations of the embedding plot (e.g., the signature names) are now automatically adjusted to prevent overlap.
Normalizing the embeddings now projects them onto the hypersphere. This is necessary for (multimodal) CorrNMF, normalizing with the sum without taking absolute values didn't make sense. It might be useful to implement multiple normalization methods in the future.